question StaticClassTest added
