--- spring_course/src/main/java/by/atmm/java/__early/initializer/ConstructorTest.java
@@ -1,11 +1,0 @@
-package by.atmm.java.__early.initializer;
-
-public class ConstructorTest {
-    public static void main(String[] args) {
-        Animal animal = new Dog();
-        //Dog dog = new Dog();
-
-        System.out.println(animal.getFullName());
-        //System.out.println(dog.getName());
-    }
-}
