+++ spring_course/src/main/java/by/atmm/java/__early/initializer/inherit/InheritTest.java
@@ -1,10 +1,11 @@
-package by.atmm.java.__early.initializer;
+package by.atmm.java.__early.initializer.inherit;
 
-public class Main {
+import by.atmm.java.__early.initializer.inherit.Inherit;
+
+public class InheritTest {
     public static void main(String[] args) {
-        //Base testBase = new Base();
+        // Base testBase = new Base();
         System.out.println("-------------------------------------------------------------------------");
         Inherit testInherite = new Inherit();
     }
-
 }
